Structural Changes (+ some bugs)
Tower package now split across three files for readability.
A few bugs fixed.
Some comments amended.
